cherry pick 27861 Add truncated_gaussian_random XPU kernel, test=kunlun (#28060)
* Add truncated_gaussian_random_op XPU kernel

* Add truncated_gaussian_random_op XPU kernel, test=kunlun

* little change, test=kunlun

* change boost_get to BOOST_GET_CONST

* change boost_get to BOOST_GET_CONST, test=kunlun

* little change, test=kunlun

* use Generator to generate random number and optimize format, test=kunlun

* little change, test=kunlun

* add TODO, test=kunlun
